Update based on PR feedback and discussion
